fix image dtype after taking mean

+++ texture.py
@@ -41,7 +41,7 @@ def haralick(image: np.ndarray, mask: np.ndarray, distance=5, mode: str = "max")
         image = np.max(image, axis=0)
         mask = np.max(mask, axis=0)
     if mode == "avg":
-        image = np.mean(image, axis=0)
+        image = np.mean(image, axis=0).astype(np.uint8)
         mask = np.max(mask, axis=0)
 
     image[~mask] = 0
